pdf_ocr_backend/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
img2text==0.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pdf2img==0.1.2
pillow==10.4.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pytesseract==0.3.13
Werkzeug==3.0.4
